+++ src/main/java/kitchenpos/menu/application/MenuProductService.java
@@ -3,7 +3,7 @@ package kitchenpos.menu.application;
 import kitchenpos.event.MenuProductValidator;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuProductRepository;
+import kitchenpos.menu.domain.repository.MenuProductRepository;
 import kitchenpos.menu.ui.dto.MenuRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
