+++ src/main/java/kitchenpos/menu/domain/repository/MenuGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.domain.repository;
 
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
