+++ src/main/java/kitchenpos/menu/domain/repository/MenuRepository.java
@@ -1,9 +1,9 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.domain.repository;
 
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
-    Long countByIdIn(List<Long> ids);
 }
