+++ src/main/java/kitchenpos/order/domain/repository/OrderLineItemRepository.java
@@ -7,9 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
-    @Query("select ol from OrderLineItem ol where ol.order.id = :orderId")
-    List<OrderLineItem> findAllByOrderId(Long orderId);
-
     @Query("select ol from OrderLineItem ol where ol.menu.id = :menuId")
     List<OrderLineItem> findAllByMenuId(Long menuId);
 }
