+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -3,8 +3,8 @@ package kitchenpos.application;
 import kitchenpos.exception.NonExistentException;
 import kitchenpos.menu.application.MenuProductService;
 import kitchenpos.menu.application.MenuService;
-import kitchenpos.menu.domain.MenuGroupRepository;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.repository.MenuGroupRepository;
+import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.ui.dto.MenuRequest;
 import kitchenpos.menu.ui.dto.MenuUpdateRequest;
 import org.junit.jupiter.api.DisplayName;
