+++ src/test/java/kitchenpos/integration/MenuProductServiceIntegrationTest.java
@@ -4,8 +4,8 @@ import kitchenpos.exception.NonExistentException;
 import kitchenpos.menu.application.MenuProductService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.repository.MenuGroupRepository;
+import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.ui.dto.MenuRequest;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
